Move data pull unit tests over to scheduled task module, delete data pull code from hrm-core
