+++ hrm-domain/scheduled-tasks/hrm-data-pull/package.json
@@ -3,7 +3,9 @@
   "version": "1.0.0",
   "description": "The easiest way to run the tests locally is to run the test DB in docker.",
   "scripts": {
-    "start": "cd dist && node ./index"
+    "start": "cd dist && node ./index",
+    "test:unit": "jest tests/unit",
+    "test": "npm run test:unit && npm run test:integration"
   },
   "dependencies" : {
     "@tech-matters/s3-client": "^1.0.0",
