+++ hrm-domain/scheduled-tasks/hrm-data-pull/tests/unit/pull-cases.test.ts
@@ -19,18 +19,18 @@ import formatISO from 'date-fns/formatISO';
 import format from 'date-fns/format';
 import addDays from 'date-fns/addDays';
 
-import * as caseApi from '../../../src/case/caseService';
-import * as context from '../../../src/data-pull-task/khp-data-pull-task/context';
-import { defaultLimitAndOffset } from '../../../src/data-pull-task/khp-data-pull-task/auto-paginate';
-import { pullCases } from '../../../src/data-pull-task/khp-data-pull-task/pull-cases';
+import * as caseApi from '@tech-matters/hrm-core/src/case/caseService';
+import * as context from '../../src/context';
+import { defaultLimitAndOffset } from '../../src/auto-paginate';
+import { pullCases } from '../../src/pull-cases';
 
 const { maxPermissions } = context;
 
 const accountSid = 'ACxxx';
 const bucket = 'docs-bucket';
 
-jest.mock('../../../src/contact/contactService');
-jest.mock('../../../src/data-pull-task/khp-data-pull-task/context');
+jest.mock('@tech-matters/hrm-core/src/contact/contactService');
+jest.mock('../../src/context');
 
 let putS3ObjectSpy = jest.fn();
 jest.mock('@tech-matters/s3-client', () => {
